2try to correct delete multiple tag delets tag

+++ sip-api/src/main/java/ch/fhnw/lst/sipapi/service/HashtagService.java
@@ -19,12 +19,11 @@ public class HashtagService {
         List<Hashtag> allHashtags = hashtagRepository.findAll();
         for (Hashtag hashtag:allHashtags) {
             hashtag.countHashtags();
-            if(hashtag.getHashtagCount()==0){
-                hashtagRepository.delete(hashtag);
-            }
+            if(hashtag.getHashtagCount()==0){hashtagRepository.delete(hashtag);}
         }
-        allHashtags.sort((h1, h2) -> Integer.compare(h2.getHashtagCount(), h1.getHashtagCount()));
-        return allHashtags;
+        List<Hashtag> correctedHashtags = hashtagRepository.findAll();
+        correctedHashtags.sort((h1, h2) -> Integer.compare(h2.getHashtagCount(), h1.getHashtagCount()));
+        return correctedHashtags;
     }
     public Hashtag findById(Long id){
         return hashtagRepository.findById(id).orElseThrow(() ->
